refactor(appsync): refactor schema-types to separate files (#9805)
refactoring the code-base for code-first schema generation because the `schema-types.ts` file was getting way too big.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*
