+++ packages/@aws-cdk/aws-appsync/lib/graphqlapi.ts
@@ -4,7 +4,7 @@ import { ManagedPolicy, Role, ServicePrincipal, Grant, IGrantable } from '@aws-c
 import { CfnResource, Construct, Duration, IResolvable, Stack } from '@aws-cdk/core';
 import { CfnApiKey, CfnGraphQLApi, CfnGraphQLSchema } from './appsync.generated';
 import { IGraphqlApi, GraphqlApiBase } from './graphqlapi-base';
-import { ObjectType, ObjectTypeProps } from './schema-types';
+import { ObjectType, ObjectTypeProps } from './schema-intermediate';
 
 /**
  * enum with all possible values for AppSync authorization type
