+++ packages/@aws-cdk/aws-appsync/lib/index.ts
@@ -4,6 +4,8 @@ export * from './key';
 export * from './data-source';
 export * from './mapping-template';
 export * from './resolver';
-export * from './schema-types';
+export * from './schema-intermediate';
+export * from './schema-field';
+export * from './schema-base';
 export * from './graphqlapi';
 export * from './graphqlapi-base';
