+++ packages/@aws-cdk/aws-appsync/lib/schema-field.ts
@@ -1,204 +1,5 @@
-/**
- * Directives for types
- *
- * i.e. @aws_iam or @aws_subscribe
- *
- * @experimental
- */
-export class Directive {
-  /**
-   * Add the @aws_iam directive
-   */
-  public static iam(): Directive{
-    return new Directive('@aws_iam');
-  }
-
-  /**
-   * Add a custom directive
-   *
-   * @param statement - the directive statement to append
-   * Note: doesn't guarantee functionality
-   */
-  public static custom(statement: string): Directive {
-    return new Directive(statement);
-  }
-
-  /**
-   * the directive statement
-   */
-  public readonly statement: string;
-
-  private constructor(statement: string) { this.statement = statement; }
-}
-
-/**
- * Properties for configuring an Intermediate Type
- *
- * @param definition - the variables and types that define this type
- * i.e. { string: GraphqlType, string: GraphqlType }
- *
- * @experimental
- */
-export interface IntermediateTypeProps {
-  /**
-   * the attributes of this type
-   */
-  readonly definition: { [key: string]: GraphqlType };
-}
-
-/**
- * Interface Types are abstract types that includes a certain set of fields
- * that other types must include if they implement the interface.
- *
- * @experimental
- */
-export class InterfaceType {
-  /**
-   * the name of this type
-   */
-  public readonly name: string;
-  /**
-   * the attributes of this type
-   */
-  public readonly definition: { [key: string]: GraphqlType };
-
-  public constructor(name: string, props: IntermediateTypeProps) {
-    this.name = name;
-    this.definition = props.definition;
-  }
-
-  /**
-   * Create an GraphQL Type representing this Intermediate Type
-   *
-   * @param options the options to configure this attribute
-   * - isList
-   * - isRequired
-   * - isRequiredList
-   */
-  public attribute(options?: BaseGraphqlTypeOptions): GraphqlType{
-    return GraphqlType.intermediate({
-      isList: options?.isList,
-      isRequired: options?.isRequired,
-      isRequiredList: options?.isRequiredList,
-      intermediateType: this,
-    });
-  }
-
-  /**
-   * Generate the string of this object type
-   */
-  public toString(): string {
-    let schemaAddition = `interface ${this.name} {\n`;
-    Object.keys(this.definition).forEach( (key) => {
-      const attribute = this.definition[key];
-      schemaAddition = `${schemaAddition}  ${key}: ${attribute.toString()}\n`;
-    });
-    return `${schemaAddition}}`;
-  }
-}
-
-/**
- * Properties for configuring an Object Type
- *
- * @param definition - the variables and types that define this type
- * i.e. { string: GraphqlType, string: GraphqlType }
- * @param interfaceTypes - the interfaces that this object type implements
- * @param directives - the directives for this object type
- *
- * @experimental
- */
-export interface ObjectTypeProps extends IntermediateTypeProps {
-  /**
-   * The Interface Types this Object Type implements
-   *
-   * @default - no interface types
-   */
-  readonly interfaceTypes?: InterfaceType[];
-  /**
-   * the directives for this object type
-   *
-   * @default - no directives
-   */
-  readonly directives?: Directive[];
-}
-
-/**
- * Object Types are types declared by you.
- *
- * @experimental
- */
-export class ObjectType extends InterfaceType {
-  /**
-   * A method to define Object Types from an interface
-   */
-  public static implementInterface(name: string, props: ObjectTypeProps): ObjectType {
-    if (!props.interfaceTypes || !props.interfaceTypes.length) {
-      throw new Error('Static function `implementInterface` requires an interfaceType to implement');
-    }
-    return new ObjectType(name, {
-      interfaceTypes: props.interfaceTypes,
-      definition: props.interfaceTypes.reduce((def, interfaceType) => {
-        return Object.assign({}, def, interfaceType.definition);
-      }, props.definition),
-      directives: props.directives,
-    });
-  }
-  /**
-   * The Interface Types this Object Type implements
-   *
-   * @default - no interface types
-   */
-  public readonly interfaceTypes?: InterfaceType[];
-  /**
-   * the directives for this object type
-   *
-   * @default - no directives
-   */
-  public readonly directives?: Directive[];
-
-  public constructor(name: string, props: ObjectTypeProps) {
-    super(name, props);
-    this.interfaceTypes = props.interfaceTypes;
-    this.directives = props.directives;
-  }
-
-  /**
-   * Generate the string of this object type
-   */
-  public toString(): string {
-    let title = this.name;
-    if(this.interfaceTypes && this.interfaceTypes.length){
-      title = `${title} implements`;
-      this.interfaceTypes.map((interfaceType) => {
-        title = `${title} ${interfaceType.name},`;
-      });
-      title = title.slice(0, -1);
-    }
-    const directives = this.generateDirectives(this.directives);
-    let schemaAddition = `type ${title} ${directives}{\n`;
-    Object.keys(this.definition).forEach( (key) => {
-      const attribute = this.definition[key];
-      schemaAddition = `${schemaAddition}  ${key}: ${attribute.toString()}\n`;
-    });
-    return `${schemaAddition}}`;
-  }
-
-  /**
-   * Utility function to generate directives
-   *
-   * @param directives the directives of a given type
-   * @param delimiter the separator betweeen directives
-   * @default - ' '
-   */
-  private generateDirectives(directives?: Directive[], delimiter?: string): string{
-    let schemaAddition = '';
-    if (!directives){ return schemaAddition; }
-    directives.map((directive) => {
-      schemaAddition = `${schemaAddition}${directive.statement}${delimiter ?? ' '}`;
-    });
-    return schemaAddition;
-  }
-}
+import { Type } from './schema-base';
+import { InterfaceType } from './schema-intermediate';
 
 /**
  * Base options for GraphQL Types
@@ -512,93 +313,4 @@ export class GraphqlType {
     type = this.isRequiredList ? `${type}!` : type;
     return type;
   }
-}
-
-/**
- * Enum containing the Types that can be used to define ObjectTypes
- */
-export enum Type {
-  /**
-   * `ID` scalar type is a unique identifier. `ID` type is serialized similar to `String`.
-   *
-   * Often used as a key for a cache and not intended to be human-readable.
-   */
-  ID = 'ID',
-  /**
-   * `String` scalar type is a free-form human-readable text.
-   */
-  STRING = 'String',
-  /**
-   * `Int` scalar type is a signed non-fractional numerical value.
-   */
-  INT = 'Int',
-  /**
-   * `Float` scalar type is a signed double-precision fractional value.
-   */
-  FLOAT = 'Float',
-  /**
-   * `Boolean` scalar type is a boolean value: true or false.
-   */
-  BOOLEAN = 'Boolean',
-
-  /**
-   * `AWSDate` scalar type represents a valid extended `ISO 8601 Date` string.
-   *
-   * In other words, accepts date strings in the form of `YYYY-MM-DD`. It accepts time zone offsets.
-   *
-   * @see https://en.wikipedia.org/wiki/ISO_8601#Calendar_dates
-   */
-  AWS_DATE = 'AWSDate',
-  /**
-   * `AWSTime` scalar type represents a valid extended `ISO 8601 Time` string.
-   *
-   * In other words, accepts date strings in the form of `hh:mm:ss.sss`. It accepts time zone offsets.
-   *
-   * @see https://en.wikipedia.org/wiki/ISO_8601#Times
-   */
-  AWS_TIME = 'AWSTime',
-  /**
-   * `AWSDateTime` scalar type represents a valid extended `ISO 8601 DateTime` string.
-   *
-   * In other words, accepts date strings in the form of `YYYY-MM-DDThh:mm:ss.sssZ`. It accepts time zone offsets.
-   *
-   * @see https://en.wikipedia.org/wiki/ISO_8601#Combined_date_and_time_representations
-   */
-  AWS_DATE_TIME = 'AWSDateTime',
-  /**
-   * `AWSTimestamp` scalar type represents the number of seconds since `1970-01-01T00:00Z`.
-   *
-   * Timestamps are serialized and deserialized as numbers.
-   */
-  AWS_TIMESTAMP = 'AWSTimestamp',
-  /**
-   * `AWSEmail` scalar type represents an email address string (i.e.`username@example.com`)
-   */
-  AWS_EMAIL = 'AWSEmail',
-  /**
-   * `AWSJson` scalar type represents a JSON string.
-   */
-  AWS_JSON = 'AWSJSON',
-  /**
-   * `AWSURL` scalar type represetns a valid URL string.
-   *
-   * URLs wihtout schemes or contain double slashes are considered invalid.
-   */
-  AWS_URL = 'AWSURL',
-  /**
-   * `AWSPhone` scalar type represents a valid phone number. Phone numbers maybe be whitespace delimited or hyphenated.
-   *
-   * The number can specify a country code at the beginning, but is not required for US phone numbers.
-   */
-  AWS_PHONE = 'AWSPhone',
-  /**
-   * `AWSIPAddress` scalar type respresents a valid `IPv4` of `IPv6` address string.
-   */
-  AWS_IP_ADDRESS = 'AWSIPAddress',
-
-  /**
-   * Type used for Intermediate Types
-   * (i.e. an interface or an object type)
-   */
-  INTERMEDIATE = 'INTERMEDIATE',
 }
